Simplest Possible WebGL program
This program is the equivalent of HelloCanvas in the WebGL Programming
Guide.  Instead of 36 lines in two files, I scrunched it down to 21
lines in one file.  The original HelloCanvas program also used a bunch
of third party libraries to abstract away the boilerplate
initialization code.  I just put the initialization code directly in my
function.

I was able to reduce the code size by getting rid of absolutely all
error handling, which you obviously wouldn’t want to do for production
code.