factored out the coffeescript into a separate file
Now that I’ve got over a hundred lines of code, and since the html has
been stable for several commits now, I split the html and coffeescript
out into separate files.  The trade off for this is the added
complexity of the development environment.  Previously I could have
written this in notepad.  Now I need a coffeescript compiler.