matrix unit tests
Added matrix unit tests and also changed the multiply method to be
arguably more coffeescript-y.  It may or may not be easier to read, but
I think the matrix multiply algorithm is a little less obscured in this
version.  Also excluded main.js from karma.  For some reason it was
trying to run it.