frustum culling first pass
Now only tessellating faces that are inside the view frustum.  The
frustum intersection test is not ideal.  If all the points of the
face’s aabb are outside the frustum, the face is assumed to be outside
the frustum.  The problem is if the frustum is inside the aabb instead,
the test gives a false negative.